Correct dependencies for JHC lib targets
Added the libs target to local-all. Previously the libs target was in an install hook which resulted
in the libs being built as the user executing "make install"
Modifications to the sources the JHC libs did not result in a recompile of the lib as expected. 
Made the JHC_LIBS targets phony. This causes them to be executed regardless of the state of the
target file. JHC appears to perform the required dependency tracking just fine.